A number guessing game using python built 8/6/2024
This is a number guessing game where the computer generates a number and the user has to guess the number if the input is lower than the number generated the computer alerts the user so and gives a hint as to what range the number lies in the max number that is in the range the computer hints the user would but 5. for example if the number was 10 then the cmputer would alert the user that the number generated is between (5,15) so it adds a number between 1 and 5 to the generated number and subtracts a number between 1 and 5 as well, giving a minimum and maximum. after seven chances it kick the user, but if the user gets it right it allows the user to continue.